+++ core/src/setup.js
@@ -1,4 +1,3 @@
-import { FastClick } from '@onsenui/fastclick';
 import './ons/platform'; // This file must be loaded before Custom Elements polyfills.
 import './polyfills/index.js';
 import './vendor/index.js';
@@ -9,28 +8,6 @@ export default function setup(ons) {
     ons._util.warn('Onsen UI is loaded more than once.');
   }
   window._onsLoaded = true;
-
-  // fastclick
-  window.addEventListener('load', () => {
-    ons.fastClick = FastClick.attach(document.body);
-
-    const supportTouchAction = 'touch-action' in document.body.style;
-
-    ons.platform._runOnActualPlatform(() => {
-      if (ons.platform.isAndroid()) {
-        // In Android4.4+, correct viewport settings can remove click delay.
-        // So disable FastClick on Android.
-        ons.fastClick.destroy();
-      } else if (ons.platform.isIOS()) {
-        if (supportTouchAction && (ons.platform.isIOSSafari() || ons.platform.isWKWebView())) {
-          // If 'touch-action' supported in iOS Safari or WKWebView, disable FastClick.
-          ons.fastClick.destroy();
-        } else {
-          // Do nothing. 'touch-action: manipulation' has no effect on UIWebView.
-        }
-      }
-    });
-  }, false);
 
   ons.ready(function() {
     ons.enableDeviceBackButtonHandler();